TIDIJ-191 Code refactoring
