+++ fleetcarpooling/lib/Models/reservation_model.dart
@@ -20,4 +20,15 @@ class Reservation {
     required this.returnTime,
     this.name,
   });
+
+  Map<String, dynamic> toMap() {
+    return {
+      'vin': vin,
+      'email': email,
+      'pickupDate': pickupDate,
+      'returnDate': returnDate,
+      'pickupTime': pickupTime,
+      'returnTime': returnTime,
+    };
+  }
 }
