+++ fleetcarpooling/lib/ReservationService/reservation_service.dart
@@ -243,62 +243,62 @@ class ReservationService implements ReservationRepository {
   }
 
   Stream<List<Reservation>> getAllReservations() {
-  DatabaseReference ref = FirebaseDatabase.instance.ref("Reservation");
-  final StreamController<List<Reservation>> controller =
-      StreamController<List<Reservation>>();
+    DatabaseReference ref = FirebaseDatabase.instance.ref("Reservation");
+    final StreamController<List<Reservation>> controller =
+        StreamController<List<Reservation>>();
 
-  ref.onValue.listen((DatabaseEvent event) async {
-    List<Reservation> userReservations = [];
+    ref.onValue.listen((DatabaseEvent event) async {
+      List<Reservation> userReservations = [];
 
-    Map<dynamic, dynamic>? values =
-        event.snapshot.value as Map<dynamic, dynamic>?;
+      Map<dynamic, dynamic>? values =
+          event.snapshot.value as Map<dynamic, dynamic>?;
 
-    if (values != null) {
-      await Future.forEach(values.entries, (entry) async {
-        try {
-          String id = entry.key;
-          String vin = entry.value['VinCar'] ?? '';
-          String email = entry.value['email'] ?? '';
+      if (values != null) {
+        await Future.forEach(values.entries, (entry) async {
+          try {
+            String id = entry.key;
+            String vin = entry.value['VinCar'] ?? '';
+            String email = entry.value['email'] ?? '';
 
-          UserRepository userRepository = UserRepository();
-          String? name = await userRepository.getUserNameByEmail(email);
+            UserRepository userRepository = UserRepository();
+            String? name = await userRepository.getUserNameByEmail(email);
 
-          DateTime pickupDate =
-              DateTime.parse(entry.value['pickupDate'] ?? '');
-          DateTime returnDate =
-              DateTime.parse(entry.value['returnDate'] ?? '');
+            DateTime pickupDate =
+                DateTime.parse(entry.value['pickupDate'] ?? '');
+            DateTime returnDate =
+                DateTime.parse(entry.value['returnDate'] ?? '');
 
-          TimeOfDay parseTimeOfDay(String time) {
-            List<int> parts = time.split(':').map(int.parse).toList();
-            return TimeOfDay(hour: parts[0], minute: parts[1]);
+            TimeOfDay parseTimeOfDay(String time) {
+              List<int> parts = time.split(':').map(int.parse).toList();
+              return TimeOfDay(hour: parts[0], minute: parts[1]);
+            }
+
+            TimeOfDay pickupTime =
+                parseTimeOfDay(entry.value['pickupTime'] ?? '');
+            TimeOfDay returnTime =
+                parseTimeOfDay(entry.value['returnTime'] ?? '');
+
+            userReservations.add(Reservation(
+              id: id,
+              vin: vin,
+              email: email,
+              pickupDate: pickupDate,
+              returnDate: returnDate,
+              pickupTime: pickupTime,
+              returnTime: returnTime,
+              name: name,
+            ));
+          } catch (e) {
+            print('Error processing reservation data: $e');
           }
+        });
+      }
 
-          TimeOfDay pickupTime =
-              parseTimeOfDay(entry.value['pickupTime'] ?? '');
-          TimeOfDay returnTime =
-              parseTimeOfDay(entry.value['returnTime'] ?? '');
+      userReservations.sort((a, b) => b.pickupDate.compareTo(a.pickupDate));
 
-          userReservations.add(Reservation(
-            id: id,
-            vin: vin,
-            email: email,
-            pickupDate: pickupDate,
-            returnDate: returnDate,
-            pickupTime: pickupTime,
-            returnTime: returnTime,
-            name: name,
-          ));
-        } catch (e) {
-          print('Error processing reservation data: $e');
-        }
-      });
-    }
+      controller.add(userReservations);
+    });
 
-    userReservations.sort((a, b) => b.pickupDate.compareTo(a.pickupDate));
-
-    controller.add(userReservations);
-  });
-
-  return controller.stream;
-}
+    return controller.stream;
+  }
 }
