+++ fleetcarpooling/lib/pages/map.dart
@@ -4,6 +4,7 @@ import 'package:fleetcarpooling/ReservationService/reservation_service.dart';
 import 'package:fleetcarpooling/VehicleManagamentService/vehicle_managament_service.dart';
 import 'package:fleetcarpooling/handlers/udp_manager.dart';
 import 'package:core/ui_elements/custom_toast.dart';
+import 'package:fleetcarpooling/ui_elements/vehicle_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:core/ui_elements/vehicle_controller.dart';
