+++ fleetcarpooling/pubspec.yaml
@@ -28,7 +28,6 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  table_calendar: ^3.0.9
   flutter:
     sdk: flutter
   intl: ^0.18.0
@@ -42,7 +41,6 @@ dependencies:
   firebase_database: ^10.3.4
   mailer: ^6.0.1
   flutter_dotenv: ^5.1.0
-  dropdown_button2: ^2.3.9
   image_picker: ^1.0.4
   firebase_storage: ^11.5.3
   google_maps_flutter: ^2.5.0
@@ -55,6 +53,7 @@ dependencies:
   firebase_messaging: ^14.7.10
   http: ^1.1.2
   cached_network_image: ^3.3.1
+  table_calendar: ^3.0.9
 
 dev_dependencies:
   flutter_test:
