+++ manuallyaddition/lib/add_vehicle_manually_form.dart
@@ -6,7 +6,6 @@ import 'package:core/ui_elements/buttons.dart';
 import 'package:core/ui_elements/colors';
 import 'package:core/ui_elements/text_field.dart';
 import 'package:core/vehicle.dart';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:dropdown_button2/dropdown_button2.dart';
