+++ manuallyaddition/pubspec.yaml
@@ -29,6 +29,10 @@ dependencies:
   cupertino_icons: ^1.0.2
   core:
     path: ../core
+  image_picker: ^1.0.5
+  firebase_storage: ^11.5.6
+  dropdown_button2: ^2.3.9
+  flutter_bloc: ^8.1.3
 dev_dependencies:
   flutter_test:
     sdk: flutter
